Just let the sqlite wrapper create the database
I'm not sure if this was a workaround for some bug, but I wasn't able to
find a mention of it online yet. And the reopen/recreate using another
library was causing issues. If you do have a startup problem, like for
instance your port is already bound by another process and the startup
bails out part of the way through, I kept getting filed that existed but
without the schema configured. Really to fix that problem I need to
refactor the startup for the activity pub database. It should be a lot
more durable around failures if it just checks (IF NOT EXISTS on schema
setup commands, and I think we'll just have to see if the user records
exist and create them if not). That way we'll always try to create the
schema we need, even if we got interrupted part of the way through.